rename cas pkg to casuploader
I want to reuse the name for the upcoming implementation that includes
both upload and download.

Leaving the existing code within the package causes a cyclic import
error between fakes, client and cas. Attempting to break the cycle
quickly grew to be too big of a change compared to this one.
