+++ go/pkg/casuploader/ioutil.go
