+++ go/pkg/tool/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/bazelbuild/remote-apis-sdks/go/pkg/tool",
     visibility = ["//visibility:public"],
     deps = [
-        "//go/pkg/cas",
+        "//go/pkg/casuploader",
         "//go/pkg/client",
         "//go/pkg/command",
         "//go/pkg/digest",
